feat: add util to transform response
